pyproject.toml
[tool.poetry]
name = "stepik-bot-app"
version = "1.0.0"
description = "Stepik Bot Web Application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "2.3.3"
gunicorn = "21.2.0"
python-telegram-bot = "20.8"
python-dotenv = "1.0.0"
requests = "2.31.0"
Werkzeug = "2.3.7"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
